Unify tab UI design across site
Reuse the tab design from the rich text editor acrsoss the site
because it is more visually consistent with USWDS.

Fixes #1912.
