+++ app/components/tabs/index.module.scss
