+++ app/routes/_gcn.docs.notices.schema.mdx
@@ -4,7 +4,7 @@ handle:
 ---
 
 import { Highlight } from '~/components/Highlight'
-import { Tab, Tabs } from '~/components/tabs/Tabs'
+import { Tab, Tabs } from '~/components/tabs'
 
 # Unified Schema
